Add files via upload
Abstract:- The Traditional manual methods for opinion analysis are time consuming and they
are limited in their ability to handle large volumes of data. Artificial intelligence
techniques, natural language processing and machine learning, have emerged as
powerful tools for automating the process of opinion analysis in social media. The
application of AI in an social media for opinion analysis. It discusses various AI
techniques that are used for sentiment analysis, emotion detection, and topic extract
and to gain insights into users opinions and attitudes. The benefits of AI-driven
opinion analysis in social media. It emphasizes the ability of AI models to process
vast amounts of data quickly, enabling real time analysis and providing valuable
insights for businesses, marketers, policy makers. AI can help identify emerging
trends, detect customer sentiment towards the products or services, and understand
public opinion on social and political issues. The aim of our project is to analyse a
solution for the perception classification at a fine grinded level, mainly the sentence
level in which polarity of sentence can be given by three categories such as positive,
negative and neutral.